+++ src/main/java/com/epam/reportportal/exception/ReportPortalException.java
@@ -17,7 +17,8 @@ package com.epam.reportportal.exception;
 
 import com.epam.ta.reportportal.ws.model.ErrorRS;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Strings;
+
+import static org.apache.commons.lang3.StringUtils.isBlank;
 
 /**
  * Base ReportPortal. Used for unclassified errors
@@ -65,7 +66,7 @@ public class ReportPortalException extends GeneralReportPortalException {
 
 	@VisibleForTesting
 	static String trimMessage(String message, int maxLength) {
-		if (Strings.isNullOrEmpty(message)) {
+		if (isBlank(message)) {
 			return "";
 		}